+++ test/resources/ns3_rebuilt.clj
@@ -1,0 +1,4 @@
+(ns resources.ns3-rebuilt
+  (:require [clojure
+             [instant :refer [read-instant-calendar]]
+             [pprint :refer [cl-format fresh-line get-pretty-writer]]]))