+++ CHANGELOG.md
@@ -12,11 +12,14 @@
 * Make `find-symbol` able to handle macros.
 
 ### Bugs fixed
+
 [clojure-emacs/clj-refactor.el#223](https://github.com/clojure-emacs/clj-refactor.el/issues/223) Fix clean-ns removing metadata when file starts with a comment.
 [#103](https://github.com/clojure-emacs/refactor-nrepl/issues/108) `find-symbol` failes in projects with `cljc` files.
 [#106](https://github.com/clojure-emacs/refactor-nrepl/issues/106) Failure to build ASTs when namespaced keywords are present in the project.
 
 ### Changes
+
+* New option for `clean-ns`, `prune-ns-form`, to avoid pruning the ns-form.
 * Get rid of the client namespace
 
 * [#118](https://github.com/clojure-emacs/refactor-nrepl/issues/118)Improve the `find-symbol` reply.  It's now a map instead of a